(shmif) prepare for safety timeout on pushfd
This replaces the calls to arcan platform fdpassing and moves them to new
variants in the shmif platform. The receive version of this returns a liveness
check function. The plan for this is to set a recvtimeout on the socket and
spin/check the liveness on timeout.

FossilOrigin-Name: 67cb1cb550cb2fbd0201af8b803fe2223c8432796ef71a36462ffa1a923eaaec