(shmif) check pending events before triggering migration
This was a fun one that was masked by the pre-refactor guard-thread shutdown
sequence. If a displayhint triggering a resize was inbound and then the DMS
pulled server-side as a forced close, the resize would check aliveness and
trigger fallback migration even though there was an _EXIT event next in the
queue.

This highlighted the same problem possibly existing with 'silent' redirection
where the server end queues a DEVICE_NODE setting the fallback connection point
then pulling the DMS. This could cause a forced migration to go to the
DEVICE_NODE set during preroll even though a newer one was intended.

To remedy both, we sweep the queue before initiating a migrate, and either
cancel it in the event of _EXIT, or update the alt_conn to match the latest
known DEVICE_NODE.

FossilOrigin-Name: e7f4222201803499489da6c7f21a6c2901a61c704beba4d0f789ccfef56a0012