(a12) regressions in upload/download
This stems from the shmif refactor mixing socket/fd when forwarding the event
to worker. There is a more subtle issue here though that comes from the state
machine handling of queued inputs/outputs versus forwarding events to the
consumer where the management pattern of client-to-client versus directory
might come and bite back.

The pending queue stepping would block forwarding the STREAMSTATUS event, which
the dir_cl is latched on for determining if it is safe to shut down or not (act
as an ACK). Now this blocking is removed, but it is nuanced if that is desired
or not.

Monitor the situation and should problems surface, consider modifying the
protocol to have explicit stream acks as part of the channel and, since it is
breaking, also take the opportunity to add an unlink.

FossilOrigin-Name: 0f2556b448fb1b2b3cc140189937cb3caae42bebb79d46112214f613bbe4ae60