(net) add match_keys, store_key to controller VM
This is only partial, we also need to route it to the database kvt in the
main process as well as forward the results back into the registered callback.

The difference for the controller is that the match_keys is asynchronous and
the scripts themselves should be responsible for caching and synching when
necessary.

The domain will also be used (later) to specify KVs that should propagate to
other servers in the message group.

FossilOrigin-Name: f24ba01f83682a9523c18a86db54f3456cd86cea8288cb4ac8bbd407829f1107