(net) move appl-runner to own thread
This is now at the stage where the store_key request is received by the
runner, so left is to coalesce and route to database and back. It also
sets the stage for on-demand spawning sources from controller.

FossilOrigin-Name: 159cf5dfe3e18e97d1e609a5754dcec7fe4d452820454708606cb78b6f44d624