(net) first bringup of launch_target
This is a fairly long journey with a few open ends still. The patch adds
placeholder support for launch_target similar to how it works for a regular
arcan_appl but from the context of the directory-server-controller.

A fundamental difference is that these pass a privilege boundary and come as
asynch calls. Now it resolves the launch parameters as per usual, but pass to
arcan-net in a mode with fixed keymaterial. It then loopback connects to the
directory as a source.

The next step is for this to be sourced in the tunnel-mode (which lacks an
option to explicitly say that the source can't support anything else really).

It goes as far as the diropen- dance is executed, and the tunnel source is
waiting for the authentication to go through, but receives no traffic.

FossilOrigin-Name: a6c4d6b6870890783aeb0fe1fa9613e26a1d4595c8a61e5a94efdc7862e2921f