(net) rename event helpers to better match function
The 100% pinning for the worker tied to joining a messaging group seem to be
associated to the refactor of shmif when acquiring a subsegment when no shm
key was passed in the event and ought to be read from the socket, but it isn't
sent anymore due to how the directory and lua processes handle the connection.

Instead of trying to band aid this, it's better to just approach the switch
to futex primitives now.

FossilOrigin-Name: bb4cdd5f60f6013d96377284db80687f7c60183b58de5de3bc1eacb5e1205c0a