(core/shmif) > BREAKING < drop the use of named primitives
There is a lot to go through with this feature as its the largest structurally
remaining item for the entire system. Currently it's in a state where it can be
worked on further, i.e. external shmif client can connect, but lots of subtle
and not so subtle caveats apply.

Checklist for pending items:

 - authoritative launch
 - subsegments
 - output segments
 - force-pushed segments
 - eventqueue killswitch
 - guardthread
 - switch from spinlock/sleep to futex
 - migration
 - memfd fallback on OpenBSD / OSX
 - arcan-net handover
 - arcan-net dirsrv uses
 - switch to _ulock on osx
 - cleanup of remaining shmif parts

FossilOrigin-Name: 2438faf4dbed0ed2556338a530ec021dca121ee217419d88ef0250a6e73734df