(core, shmif) authoritative launch fixed, prepare for >1 fdpassing
Strange one - from yesterday's tests it seemed that creating a socketpair,
sending the shm descriptor to it and then closing the client side in the parent
always caused ECONNRESET when trying to receive the descriptor agian.

Now this still happens in the fork+exec case, but not so in the subsegment one
which is something to look out for. Still we prepare for changing NEWSEGMENT to
actually receive both memory and socket in the event handler rather than socket
in event and then get memory from the socket.

This change also adds a SHMIF_SOCKIN_MEMFD to pair with SHMIF_SOCKIN_FD when we
configure the environment, and to check for that when we acquire an inherited
connection.

FossilOrigin-Name: 4c134bb64c0e6ef0df3341831054e3ea657afa080cf0980b0c5e01d9adad3cd6