(shmif/net) close serveral descriptor leaks from the refactor
The biggest gotcha was that shmifsrv, regardless of source, always ran through
the authentication string (as that was previously for finding the named
primitives) even if the connection was inherited through calling spawn_client.

This caused the shmfd to be passed over the connection socket and then be used
for the first BCHUNK_IN. This was first visible in the directory server
spawning a worker as that provided the initial directory list.

Left for the refactor now is to ensure that the lua-appl-runner and, what
started this ordeal, the dynamic source launching, are set up correctly since
that is a mediated-proxy-spawn shmif connection which was racy in the previous
model.

Then we can bring in the futexes, ensure that it works on various BSDs (memfd
being a problem) and add the OSX specific __ulock calls.

FossilOrigin-Name: 1fa8359b2cafd6335ad1ac4f4ab920a4029b13ddb066eae33dd91587429dc2e8