(shmif) migrate - scan_device_node event didn't increment
This caused arcan-net in dynamic launch mode to livelock, but the bigger
concern is >why< it decided to migrate in the first place. It could be due to
DEVICE_NODE being sent from the desktop and passed unchecked which is an
interesting problem.

The open question (and likely necessary as an option to launch_target) is if we
should allow a migration or not, and what the mechanics would be for that.
Since DEVICE_NODE are in the local namespace or in the a12s:// keyspace we'd
need to inject a keystore with the valid targets, but with no enumeration
available to the sink it needs some regulation from the controller.

The safest option is to simply re-attach it as a source and let the control
channel regulate all such events for the time being.

FossilOrigin-Name: 0ea5c4b95f612fc8e82aa9ec7ba31f590a22f1ed241e5e048d21469876b5aa99