(net) unbreak joining appl-ctrl from worker
This was one of (hopefully only one or two more) the breaks from switching to
named primitives. Previously we sent the one socket to the new controller
process and the other to the worker, then that was used to initiate a subseq
like request. The problem with that is that the refactor now requires both
socket and shm- descriptor in the NEWSEGMENT request.

To work around this we fake connecting a primary through a socket as if it was
an inherited spawn, but not setting the descriptor number for the shmpage
forcing it to think it's a named connection.

FossilOrigin-Name: 5da424cacdb0f451a22f8a3c27b8b7e7448fe73d99f91fc45f380531cb4e4238