(net) expose [autorun] table in config.lua for directory server
The point of this is to have 'always on' controller scripts that can spawn
global dynamic sources without having someone actually join the messaging group
and launching on demand. We still require an appl for a corresponding
controller to be present (but not the other way around), that restriction might
be dropped in the future.

FossilOrigin-Name: d253bd54c945ead773b62894625b8ba6182a7a82c97124cc5f4c80ee6b30a0c6