(net) mark state-failures with reason, propagate on last-words
Getting errors across from why a directory-link failed was annoying as the
log-trace still isn't getting propagated and we needed a short message for
the event handler to make key-failures more noticeable.

Dynamic changes should now be enumerated from the directory link stage, only
synch and .appl / .appl-controller request todo there. Then we need to be able
to specify which level of synch the linked directory should have. The issue
there is who is the authority on which .appl it is that should be distributed.
Simply having a timestamp and sending the latest isn't viable as one might want
to rollback when a deployment is broken.

FossilOrigin-Name: f06db63b9e964a2dc45ee2562c85e956f8903cb689d5c7fb748d6f7707fc37c6