(a12) rework of logging and tunneling
This was a heavy one, first rework so that key authentication functions and
others more consistently pass the a12_state around as it will be needed for
attaching auxiliary keys later.

Then add a tracetag member to the a12_state and add it to all a12int_trace
related calls, this had some ugly macroing to be workable since it is used
allover.

Lastly, rework so that we can start to handle multiple tunnels. The linked
directory server will need this shortly. This means that each worker that pokes
the tunnels gets a thread and a signal pipe to the ioloop to notify when there
is outbound data to be forwarded from a12_write_tunnel.

FossilOrigin-Name: 9572f6b81873d145f442a19965156dd869c45f93a0c5dd8568acaaebed731965