(fsrv-net) swap IN/OUT before passing to a12
This was another source of the upload / download headache. Events coming from
open_nonblock with nowrite comes as OUT into the frameserver. Thus the
frameserver is to provide data. That means the event forwarded to a12 should be
IN as it is expecting input.

FossilOrigin-Name: 5693406013e525eb81950818a7c96af737d180e5318be7af6b695c441791a071