(net) clean up identifier use between main<->work<->ctrl
This was a low priority fixme until it became a blocker for
launch_target(clid). Now the worker to ctrl identifier sent in NETSTATE is
H(Kpub | applname) that is used for server-side state persistence.

To do that the applname is sent upon parent permission check (as the worker
doesn't track, only forward the IDENT event when joining). At the same time,
the client provided IDENT petname (which was used as prefix for MESSAGE when in
ctrl-less broadcast mode) will be forwarded to ctrl process after resolving
collision in order to be used for reverse requests where mapping from a ctrl-
local numerical ID to the corresponding struct dircl necessary for parent to
find the worker process.

FossilOrigin-Name: ac5ebe725447dfe6362a8cc01c88e92ed550eec1899968a51c2342552042aa8e