(net) add option to mask source from being announced
This is another step towards directed launch_target, a mask is set alongside
the ephemeral key the server-side source uses. Before announcing a new source,
check so that the client attached to a worker is eligible (we'll also need to
ensure that for diropen request, as well as a netstate type for a mask that
applies to a single client, and drop such client-source pairs on leave).

FossilOrigin-Name: 2b287b13b54af06a65cba295f43688c3dbdae2587d77b0d9f356768044965355