(net) prepare for appl-selfhost (server-side arcan_lwa)
This unifies the process spawning from launch_target to a function that can be
used elsewhere for server-side launched sources. The main purpose for this is
to let (with permission) a client source an appl that is ran server-side as a
source for when the full arcan-stack isn't available locally but the smash
viewer is.

FossilOrigin-Name: 25b9cd524ea3fec58357359a0079ddec22b918b2332b7eb3a898cccad263c4e8