(net) push-new-appl working
These do not get persisted by default, so a server restart would lose the
entry. Pending the move to db-backed appl+ctrl storage, this is fine as it is
just to support a quicker developer workflow. Persisting should likely be a
function in config.lua space, with an option to autopersist with version
history and the option to rollback.

FossilOrigin-Name: 3a82979783d2497c26edb20ec084546e90b01c55e6ef8c4d51074077ffb85d24