(net) add helpers for record/replay troubleshooting
There is an issue with --host-appl that causes the tunneled state machine to
randomly enter an invalid state specifically between command boundaries.
Hosting the same runner/appl without the directory wrapper is fine. Since the
MAC for the outer packet is correct, it should be something in the path of
reading/writing to/from the tunnel.

To help capture things like that, this adds a _DEBUG only entrypoint for
recording the decrypted pre-authenticated bytestream, and another for later
replaying it (disabling decryption and MAC comparison).

FossilOrigin-Name: b04a3ec04af1297e00557cedd2861f36dff1349f73ab7560656fa9300911d117