(net) dir-cl exec handling cleanup, 100% in dir_srv worker fix
The worker one might need some further consideration, the ioloop only handled
userfd2 if there was input - not if there was an error. Digging in the logs
didn't show a reason for that explicit check.

The only target relying on userfd2 is the dir_srv_worker that rotates with
a12_btransfer_outfd (as unpack can rotate which file is next in queue).

A possible reason would be reading from a file that both triggers something
like EPIPE/HUP while at the same time having more data to read.

For the dir_cl part it was a previous SIG_CHLD ignore that would cause the
waitpid to never get the status. Now we set it each time we exec but there
are more signal fsckery when it comes to debugging (monitor being a likely
cause of the regression) and hot-reload. The later will get more testing
as we progress with the consort project.

FossilOrigin-Name: affacd8d737e6a44a5e6cb5d3aee746f4417a2a30996549dc7bef21392f9f81b