(net) appl_cleanup: try to increase success rate
We only aggressively clean up in the extraction dir applname subfolder as
whatever else other dependencies might conjure up is problematic in the way it
is done now (all of this need to become config.lua like options).

In the process, if openal- pulls in dep on pulseaudio during probing (which
shouldn't happen in the lwa- build but still does, new rebase on more current
master is needed) it will leave a dangling empty pulse folder (because it
doesn't actually do a pulse audio connection ...) and that breaks tracking of
what was extracted or not.

Just unlinkat("pulse") for the time being.

FossilOrigin-Name: 1e395360f5ec860f94e3189e3d41d4df0f105f53771674e245846d2ca348afa3