(lua) extend open_nonblock with table form
This is in preparation for handling parallel transfers through
afsrv_net/arcan-net. The problem this is meant to solve is that if an appl
does something like:

    stdin = net_open("@stdin", ...)
		decode = launch_decode(...)
    open_nonblock(stdin, false, "appl:/something", decode)

Any subsequent transfers will be queued and not dispatched until decode
finishes streaming. This breaks simple cases like an a12-ctrl+appl that may
stream video from a static source while also presenting static imagery that
comes server-side.

There's no reasonable heuristic for this since afsrv_decode etc. can be used
for static imagery as well, the option is either to instruct _decode to locally
cache before decoding (which can be gigabytes of data and still block) or,
as the approach is here, to add hinting:

    open_nonblock(stdin, {parallel = true}, "appl:/something", decode)

Then (next patch) let arcan-net see that this should be parallelised and
request it on a different channel. This gives the appl the option of ~255
(assuming no other dynamic sourcing or tunneling) ongoing transfers.

FossilOrigin-Name: 75fde94de0ce61ec6a8f313e38dd88fe68b64e85adad15266997acff5d4f4283