(net) expose debug report generation as .report resource
This solves for half of the triaging point. It makes it so a developer
permission key can generate and download the report for the appl it specifies,
a config option to flush after this has occured (the flush happens instantly,
meaning that if the connection is severed while the report is downloaded, it is
lost).

It can be parried for by tracking the transfer and waiting for STREAMSTATUS to
latch the unlink, if needed. It's just fringe enough to not warrant the effort
right now.

It also adds config.lua scope controls to block the flush, and admin functions
to enumerate installed appls and explicitly flush.

What's missing is the .ctrl layer, both for merging into the report itself, and
more importantly - an external ticket service config and functions to issue a
ticket into it.

FossilOrigin-Name: 66835228e0347ec4ae577e446486b9ab6c145e12ac383ff18a363427c0eef87a