(net) [regression warning] refactor -l -- /shmif/client hosting
This will break pull use of arcan-net -- /some/bin.

To get away from the fork() + continued exec use in arcan-net which stops
thread sanitizer from working, and make net.c less messy -- we split into a new
binary, arcan-net-session.

This binary will take care of spinning up the shmif client and mapping to an
a12 sink, but also prepare for two major features and a tui UI for runtime
parameter controls.

The first feature is to be able to resume an interrupted connection by
leveraging the crash recovery _RESET mechanism. This is done by tracking the
public-key used to authenticate, and the status of the shmif client primary
when the a12 connection has terminated. If the client is still alive, bind it
to the public-key and look for that on new connections.

The second feature is a bit more complex, but it is to have a broadcast option
where encoded video frames are kept from the session that the first source
gets.

Any new connections are then fed these frames within a GOP-sequence friendly
window as passthrough. This lets one a12 connection 'drive' the shmif client,
and the others view it. It will be single window only (and uncomposited, if
composition is wanted, host through lwa+wm).

FossilOrigin-Name: 7e0e57c14bf8f554d988fe007b36c1b2b5b4e90151de9ff5d05a2ff1c273bb41