(net) send reset/register on resumption
This works with the caveat that if the sink side WM deletes the window, causing
the connection to terminate, this will not cause the termination to hit and the
old session will still recover.

Similarly, if say, afsrv_terminal: #exit kills the window the server will still
try to resumption it. Both suggest that there is an issue with lifecycle
tracking at these transitions.

FossilOrigin-Name: 78c8cffa116f565e0b5e240423fc173d247edb18f7666f796b53943fbc835509