(net) session: wire in first attempt at frame cache for casting
This is for the last big step in the feature matrix of what we need as building
blocks for the intended first round of a12 appls, i.e. letting multiple sinks
view the same source. There is a lot that can be done with it, but settle for
just pushing towards PoC stage by adding a mode to session where the first
connected sink becomes the 'driver' and the frames that are produced gets
mirrored to any other sink that connects.

The main complications to look out for will be the usual 'how to handle
secondary surfaces' in order to preserve deferred composition, with cat9 open
embed style being the tough example.

FossilOrigin-Name: 8f1c2243bbfc95530dbbea749da75bc0ee576b6e5f6ecb38c5e9ed4117b0454e